update build.bat
As the batch file itself says, this is not expected to work, in
general. Rather, I (Jamis) use it when building a new windows
release, and so I keep it up to date with my current windows
configuration.

If someone wants to work on a more general build environment
for windows, I'll be happy to include it.