Adds missing helper methods for pragmas.
Updated the methods for pragma calls according to the currently
supported pragmas listed on https://sqlite.org/pragma.html. The previous
behaviour is maintained and extended by the new pragmas.
Only thing changed is the auto_vacuum pragma, which has changed in the
commit from 2007-04-26:
https://www.sqlite.org/cgi/src/info/f6a6d2b8872c05089810b1e095f39011f3035408

So, the minimal version for this to work is SQLite 3.4.0