Add SQLite3::Database#extended_result_codes=
If set to true, this enables extended result codes for the database,
which provide more information about separate failure types, such
as which type of constraint has been violated.

Support for this was added in SQLite 3.3.8, which was released over
10 years ago.  I didn't add extconf checks for this because it was
released so long ago, but if SQLite versions less than 3.3.8 are
supported, it's fairly easy to do that.  I'm not sure what behavior
is desired in that case.  We could either not define the method,
define it and have it always raise an exception, or define it and
have it do nothing.