define_aggregator respects step's arity again
This has been undocumented behavior before: while create_aggregate and
create_aggregate_handler effectively used an arity of -1, the library
user could actually define functions with a different arity by calling
define_aggregator directly. I was not aware of this behavior and
therefore broke it by forcing the arity to -1. Now I added the required
shim and a test.